+++ .github/workflows/main.yml
@@ -64,10 +64,10 @@ jobs:
         uses: actions/checkout@v3
     
       - name: Migrate Up
-        run: migrate -path=./api/db/migrations/postgres/ -database postgres://postgres:postgres@postgres-db:5432/guide-system?sslmode=disable up 
+        run: migrate -path=./api/provider/db/migrations/postgres/ -database postgres://postgres:postgres@postgres-db:5432/guide-system?sslmode=disable up 
 
       - name: Migrate Down
-        run: migrate -path=./api/db/migrations/postgres/ -database postgres://postgres:postgres@postgres-db:5432/guide-system?sslmode=disable down -all
+        run: migrate -path=./api/provider/db/migrations/postgres/ -database postgres://postgres:postgres@postgres-db:5432/guide-system?sslmode=disable down -all
 
   push-ui-container:
     needs: [build-ui]
