+++ api/repository/repository.go
@@ -1,6 +1,6 @@
 package repository
 
-import "github.com/st-matskevich/audio-guide-bot/api/db"
+import "github.com/st-matskevich/audio-guide-bot/api/provider/db"
 
 type Repository struct {
 	DBProvider db.DBProvider
